# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yapaia Go Core API",
    "version": "0.0.1",
    "description": "Automatisch generiert aus den tatsächlich registrierten Fastify-Routen (apps/core/src/index.ts, onRouteHook) und den @yapaia/shared-JSON-Schemas (packages/shared/src/schemas/). Quelle der Wahrheit ist der Code, nicht dieses Dokument -- bei Verdacht auf Abweichung: `pnpm openapi:check` (CI-Gate \"Spec aktuell\") bzw. docs/03-api-spec.md für den erzählenden Überblick. WebSocket (/ws/v1) und MQTT sind hier NICHT modelliert (OpenAPI eignet sich dafür nicht) -- siehe docs/03-api-spec.md §3/§4."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Core-API (relative Pfade -- funktioniert unverändert hinter HA-Ingress, W-15)"
    }
  ],
...
  "paths": {
...
    "/api/v1/navigation/destination": {
      "post": {
        "summary": "Convenience für HA/Add-ons: geocodet, routet, startet optional (`autostart`)",
        "tags": [
          "Navigation"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NavState"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "summary": "Geocoding (Photon → Fallback), bias auf Position",
        "tags": [
          "Suche"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResult"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/reverse": {
      "get": {
        "summary": "Reverse-Geocoding",
        "tags": [
          "Suche"
        ],
        "responses": {
          "200": {
            "description": "Erfolgreiche Antwort",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResult"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Fehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NavState": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "routing",
              "navigating",
              "paused",
              "arrived",
              "off_route"
            ],
            "description": "Navigation state"
          },
          "route_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "ID of current route or null"
          },
          "next_maneuver": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "index": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Maneuver index in route"
                  },
                  "type": {
                    "type": "string",
                    "description": "Maneuver type (Valhalla enum, e.g. \"turn_left\", \"roundabout_enter\")"
                  },
                  "instruction": {
                    "type": "string",
                    "description": "Localized instruction text"
                  },
                  "street_names": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Names of streets involved"
                  },
                  "distance_m": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Length of maneuver segment in meters"
                  },
                  "begin_shape_index": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Start index in route geometry"
                  },
                  "lanes": {
                    "type": [
                      "array",
                      "null"
                    ],
                    "items": {
                      "type": "object",
                      "properties": {
                        "lane_index": {
                          "type": "integer",
                          "minimum": 0,
                          "description": "Lane index from left"
                        },
                        "is_usable": {
                          "type": "boolean",
                          "description": "Whether vehicle can use this lane"
                        },
                        "direction": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Direction indicator if available"
                        }
                      },
                      "required": [
                        "lane_index",
                        "is_usable"
                      ],
                      "additionalProperties": false
                    },
                    "description": "Optional lane information"
                  },
                  "duration_s": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Planned duration of this maneuver segment in seconds (Valhalla time), optional (E04-T2)"
                  }
                },
                "required": [
                  "index",
                  "type",
                  "instruction",
                  "street_names",
                  "distance_m",
                  "begin_shape_index"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ],
            "description": "Next upcoming maneuver or null"
          },
          "distance_to_maneuver_m": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "description": "Distance to next maneuver in meters or null"
          },
          "distance_remaining_m": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "description": "Total distance remaining to destination or null"
          },
          "duration_remaining_s": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "description": "Estimated time to destination in seconds or null"
          },
          "eta": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Estimated time of arrival, UTC ISO 8601 with offset (client formats locally, W-22), or null"
          },
          "speed_kmh": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "description": "Current speed in km/h or null"
          },
          "speed_limit_kmh": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 5,
            "maximum": 130,
            "description": "Speed limit from map data (5–130 or null for unknown)"
          },
          "altitude_m": {
            "type": [
              "number",
              "null"
            ],
            "description": "Current altitude in meters above mean sea level or null"
          },
          "destination": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "latlng": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90,
                        "description": "Latitude in degrees (-90 to 90)"
                      },
                      "lon": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180,
                        "description": "Longitude in degrees (-180 to 180)"
                      }
                    },
                    "required": [
                      "lat",
                      "lon"
                    ],
                    "additionalProperties": false
                  },
                  "name": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "Destination name or null"
                  }
                },
                "required": [
                  "latlng",
                  "name"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ],
            "description": "Destination with coordinates and optional name or null"
          }
        },
        "required": [
          "status",
          "route_id",
          "next_maneuver",
          "distance_to_maneuver_m",
          "distance_remaining_m",
          "duration_remaining_s",
          "eta",
          "speed_kmh",
          "speed_limit_kmh",
          "altitude_m",
          "destination"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code identifier"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": [
                  "object",
                  "null"
                ],
                "description": "Optional additional error context"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false,
            "description": "Error object with code, message, and optional details"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Short name of the result (e.g. \"Vaduz\")"
          },
          "label": {
            "type": "string",
            "description": "Full display label (e.g. \"Vaduz, Liechtenstein\")"
          },
          "latlng": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number",
                "minimum": -90,
                "maximum": 90,
                "description": "Latitude in degrees (-90 to 90)"
              },
              "lon": {
                "type": "number",
                "minimum": -180,
                "maximum": 180,
                "description": "Longitude in degrees (-180 to 180)"
              }
            },
            "required": [
              "lat",
              "lon"
            ],
            "additionalProperties": false
          },
          "type": {
            "type": "string",
            "description": "Result category (e.g. \"city\", \"street\", \"coordinates\")"
          },
          "source": {
            "type": "string",
            "enum": [
              "photon",
              "nominatim",
              "coords",
              "lite"
            ],
            "description": "Backend that produced this result; \"lite\" = offline SQLite/FTS5 fallback used when Photon is down/disabled (E05-T5, W-12)"
          },
          "out_of_coverage": {
            "type": "boolean",
            "description": "True if the result lies outside all installed map regions (Vorgriff W-09)"
          }
        },
        "required": [
          "name",
          "label",
          "latlng",
          "type",
          "source"
        ],
        "additionalProperties": false
      },
...
    }
  }
}